requirements.txt
streamlit>=1.20.0
moviepy>=1.0.3
Pillow>=9.0.0
numpy>=1.21.0